# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Visit Lab",
        "contact": {
            "email": "saviogmoiagaia@gmail.com"
        },
        "version": "3.0.0"
    },
    "paths": {
        "/api/laboratories": {
            "get": {
                "tags": [
                    "Laboratorio"
                ],
                "summary": "Retorna os laboratorios. Usuario deve ser um professor ou administrador. Se for um professor, vai retornar apenas os laboratórios que ele é coordenador",
                "operationId": "4622a11d2d127d35109f12324df4e1d3",
                "responses": {
                    "200": {
                        "description": "Retorna os laboratórios"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/laboratories/": {
            "post": {
                "tags": [
                    "Laboratorio"
                ],
                "summary": "Cria um laboratório",
                "description": "Cria um laboratorio",
                "operationId": "cc14cd4e39974fa28397a4c09f80513d",
                "requestBody": {
                    "description": "Json informações necessárias",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "local"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "nome do laboratório",
                                        "type": "string"
                                    },
                                    "local": {
                                        "description": "Local com predio e sala do laboratorio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Login não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": [],
                        "0": {
                            "api_key": []
                        }
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/laboratories/{laboratory}": {
            "get": {
                "tags": [
                    "Laboratorio"
                ],
                "summary": "Obtém um laboratorio pelo id",
                "operationId": "5b1c536adce5b2a3f6e9edd9d335d74b",
                "parameters": [
                    {
                        "name": "laboratory",
                        "in": "path",
                        "description": "Id do laboratório",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pega um laboratorio"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Laboratorio"
                ],
                "summary": "Edita um laboratório",
                "description": "Atualiza um laboratorio",
                "operationId": "ce44ff2b2ecb3232d3ea92cb8b4d4dfa",
                "parameters": [
                    {
                        "name": "laboratory",
                        "in": "path",
                        "description": "Id do laboratório",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Json informações necessárias",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "local"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "nome do laboratório",
                                        "type": "string"
                                    },
                                    "local": {
                                        "description": "Local com predio e sala do laboratorio",
                                        "type": "string"
                                    },
                                    "login": {
                                        "description": "Login pra qual será criado uma solicitação para o usuário ser coordenador",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Login não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Laboratorio"
                ],
                "summary": "Deleta um laboratorio pelo id",
                "operationId": "81964c7b2bfeff8205cc4eee3526de0b",
                "parameters": [
                    {
                        "name": "laboratory",
                        "in": "path",
                        "description": "Id do laboratório",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laboratorio deletado"
                    },
                    "400": {
                        "description": "Erro"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Access token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "description": "Chave de aplicação para autenticação na API",
                "name": "X-Application-Key",
                "in": "header"
            }
        }
    }
}